Add files via upload
In this project, I created a word-guessing game called Snowman. Roughly speaking, if you guess a wrong letter, it will create a small part of the snowman and you will be lost until it gets over the appearance of 8 parts. If you finish all of the guessed letter before getting over 8 parts of snowman, you will win.